Refactor custom field, add config/install ymls, add test-controller and make some other minor edits
