--- config/install/dolebas_payments.stripeconfig.yml
@@ -1,1 +1,0 @@
-dolebas_payments:
