+++ config/install/field.field.node.dolebas_transaction.field_amount.yml
@@ -1,0 +1,22 @@
+langcode: en
+status: true
+dependencies:
+  config:
+    - field.storage.node.field_amount
+    - node.type.dolebas_transaction
+id: node.dolebas_transaction.field_amount
+field_name: field_amount
+entity_type: node
+bundle: dolebas_transaction
+label: Amount
+description: ''
+required: false
+translatable: false
+default_value: {  }
+default_value_callback: ''
+settings:
+  min: null
+  max: null
+  prefix: ''
+  suffix: ''
+field_type: integer
