+++ config/install/field.field.node.dolebas_transaction.field_currency.yml
@@ -1,0 +1,18 @@
+langcode: en
+status: true
+dependencies:
+  config:
+    - field.storage.node.field_currency
+    - node.type.dolebas_transaction
+id: node.dolebas_transaction.field_currency
+field_name: field_currency
+entity_type: node
+bundle: dolebas_transaction
+label: Currency
+description: ''
+required: false
+translatable: false
+default_value: {  }
+default_value_callback: ''
+settings: {  }
+field_type: string
