+++ config/install/field.field.node.dolebas_transaction.field_stripe_token.yml
@@ -1,0 +1,18 @@
+langcode: en
+status: true
+dependencies:
+  config:
+    - field.storage.node.field_stripe_token
+    - node.type.dolebas_transaction
+id: node.dolebas_transaction.field_stripe_token
+field_name: field_stripe_token
+entity_type: node
+bundle: dolebas_transaction
+label: 'Stripe Token'
+description: ''
+required: false
+translatable: false
+default_value: {  }
+default_value_callback: ''
+settings: {  }
+field_type: string
