+++ config/install/field.storage.node.field_dolebas_transaction.yml
@@ -1,0 +1,21 @@
+langcode: en
+status: true
+dependencies:
+  module:
+    - dolebas_payments
+    - node
+id: node.field_dolebas_transaction
+field_name: field_dolebas_transaction
+entity_type: node
+type: dolebas_transaction_field_type
+settings:
+  max_length: 255
+  is_ascii: false
+  case_sensitive: false
+module: dolebas_payments
+locked: false
+cardinality: 1
+translatable: true
+indexes: {  }
+persist_with_no_fields: false
+custom_storage: false
